requirements.txt
appdirs==1.4.4
attrs==25.4.0
banal==1.0.6
blinker==1.9.0
cattrs==25.3.0
certifi==2025.10.5
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
csv-reconcile==0.3.2
cycler==0.12.1
Cython==0.29.37
Flask==2.3.3
Flask-Cors==3.0.10
fonttools==4.60.1
geojson-rewind==1.2.1
geomet==1.1.0
geopandas==1.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mgrs==1.5.0
normality==2.6.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pygbif==0.6.5
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-cache==1.2.1
shapely==2.1.2
six==1.17.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
Werkzeug==3.1.3
